+++ src/OpenTelemetry/BaseExporter.cs
@@ -46,6 +46,11 @@ namespace OpenTelemetry
         private int shutdownCount;
 
         /// <summary>
+        /// Gets the parent <see cref="BaseProvider"/>.
+        /// </summary>
+        public BaseProvider ParentProvider { get; internal set; }
+
+        /// <summary>
         /// Exports a batch of telemetry objects.
         /// </summary>
         /// <param name="batch">Batch of telemetry objects to export.</param>
@@ -74,7 +79,7 @@ namespace OpenTelemetry
         {
             if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
             {
-                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds));
+                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative.");
             }
 
             if (Interlocked.Increment(ref this.shutdownCount) > 1)