+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/IncomingRequestsCollectionsIsAccordingToTheSpecTests.cs
@@ -27,8 +27,10 @@ using Moq;
 using OpenTelemetry.Trace;
 #if NETCOREAPP2_1
 using TestApp.AspNetCore._2._1;
-#else
+#elif NETCOREAPP3_1
 using TestApp.AspNetCore._3._1;
+#else
+using TestApp.AspNetCore._5._0;
 #endif
 using Xunit;
 
@@ -86,7 +88,7 @@ namespace OpenTelemetry.Instrumentation.AspNetCore.Tests
 
                 for (var i = 0; i < 10; i++)
                 {
-                    if (processor.Invocations.Count == 2)
+                    if (processor.Invocations.Count == 3)
                     {
                         break;
                     }
@@ -98,8 +100,8 @@ namespace OpenTelemetry.Instrumentation.AspNetCore.Tests
                 }
             }
 
-            Assert.Equal(2, processor.Invocations.Count); // begin and end was called
-            var activity = (Activity)processor.Invocations[1].Arguments[0];
+            Assert.Equal(3, processor.Invocations.Count); // SetParentProvider/Begin/End called
+            var activity = (Activity)processor.Invocations[2].Arguments[0];
 
             Assert.Equal(ActivityKind.Server, activity.Kind);
             Assert.Equal("localhost", activity.GetTagValue(SemanticConventions.AttributeHttpHost));