+++ test/OpenTelemetry.Instrumentation.Http.Tests/HttpWebRequestTests.Basic.netfx.cs
@@ -78,8 +78,8 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
 
             using var response = await request.GetResponseAsync();
 
-            Assert.Equal(2, activityProcessor.Invocations.Count); // begin and end was called
-            var activity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+            Assert.Equal(3, activityProcessor.Invocations.Count);  // SetParentProvider/Begin/End called
+            var activity = (Activity)activityProcessor.Invocations[2].Arguments[0];
 
             Assert.Equal(parent.TraceId, activity.Context.TraceId);
             Assert.Equal(parent.SpanId, activity.ParentSpanId);
@@ -96,6 +96,53 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
         }
 
         [Fact]
+        public async Task HttpWebRequestInstrumentationInjectsHeadersAsyncWhenActivityIsNotRecorded()
+        {
+            ActivityContext contentFromPropagator = default;
+            var activityProcessor = new Mock<BaseProcessor<Activity>>();
+            var propagator = new Mock<TextMapPropagator>();
+            propagator.Setup(m => m.Inject(It.IsAny<PropagationContext>(), It.IsAny<HttpWebRequest>(), It.IsAny<Action<HttpWebRequest, string, string>>()))
+                .Callback<PropagationContext, HttpWebRequest, Action<HttpWebRequest, string, string>>((context, message, action) =>
+                {
+                    contentFromPropagator = context.ActivityContext;
+                });
+
+            Sdk.SetDefaultTextMapPropagator(propagator.Object);
+            using var shutdownSignal = Sdk.CreateTracerProviderBuilder()
+                .AddProcessor(activityProcessor.Object)
+                .AddHttpWebRequestInstrumentation()
+                .Build();
+
+            var request = (HttpWebRequest)WebRequest.Create(this.url);
+
+            request.Method = "GET";
+
+            var parent = new Activity("parent")
+                .SetIdFormat(ActivityIdFormat.W3C)
+                .Start();
+            parent.TraceStateString = "k1=v1,k2=v2";
+            parent.ActivityTraceFlags = ActivityTraceFlags.None;
+
+            using var response = await request.GetResponseAsync();
+
+            // By default parentbasedsampler is used.
+            // In this case, the parent is the manually created parentactivity, which will have TraceFlags as None.
+            // This causes child to be not created.
+            Assert.Equal(1, activityProcessor.Invocations.Count);
+
+            Assert.Equal(parent.TraceId, contentFromPropagator.TraceId);
+            Assert.Equal(parent.SpanId, contentFromPropagator.SpanId);
+            Assert.NotEqual(default, contentFromPropagator.SpanId);
+
+            parent.Stop();
+            Sdk.SetDefaultTextMapPropagator(new CompositeTextMapPropagator(new TextMapPropagator[]
+            {
+                new TraceContextPropagator(),
+                new BaggagePropagator(),
+            }));
+        }
+
+        [Fact]
         public async Task HttpWebRequestInstrumentationInjectsHeadersAsync_CustomFormat()
         {
             var propagator = new Mock<TextMapPropagator>();
@@ -107,9 +154,10 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
                 });
 
             var activityProcessor = new Mock<BaseProcessor<Activity>>();
+            Sdk.SetDefaultTextMapPropagator(propagator.Object);
             using var shutdownSignal = Sdk.CreateTracerProviderBuilder()
                 .AddProcessor(activityProcessor.Object)
-                .AddHttpWebRequestInstrumentation(options => options.Propagator = propagator.Object)
+                .AddHttpWebRequestInstrumentation()
                 .Build();
 
             var request = (HttpWebRequest)WebRequest.Create(this.url);
@@ -124,9 +172,9 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
 
             using var response = await request.GetResponseAsync();
 
-            Assert.Equal(2, activityProcessor.Invocations.Count); // begin and end was called
+            Assert.Equal(3, activityProcessor.Invocations.Count); // SetParentProvider/Begin/End called
 
-            var activity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+            var activity = (Activity)activityProcessor.Invocations[2].Arguments[0];
 
             Assert.Equal(parent.TraceId, activity.Context.TraceId);
             Assert.Equal(parent.SpanId, activity.ParentSpanId);
@@ -140,6 +188,11 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
             Assert.Equal("k1=v1,k2=v2", tracestate);
 
             parent.Stop();
+            Sdk.SetDefaultTextMapPropagator(new CompositeTextMapPropagator(new TextMapPropagator[]
+            {
+                new TraceContextPropagator(),
+                new BaggagePropagator(),
+            }));
         }
 
         [Fact]
@@ -162,7 +215,7 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
             using var c = new HttpClient();
             await c.SendAsync(request);
 
-            Assert.Equal(0, activityProcessor.Invocations.Count);
+            Assert.Equal(1, activityProcessor.Invocations.Count);
         }
 
         [Fact]
@@ -178,7 +231,7 @@ namespace OpenTelemetry.Instrumentation.Http.Tests
             using var c = new HttpClient();
             await c.GetAsync(this.url);
 
-            Assert.Equal(0, activityProcessor.Invocations.Count);
+            Assert.Equal(1, activityProcessor.Invocations.Count);
         }
 
         [Fact]