+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlEventSourceTests.netfx.cs
@@ -128,9 +128,9 @@ namespace OpenTelemetry.Instrumentation.SqlClient.Tests
 
             fakeSqlEventSource.WriteEndExecuteEvent(objectId, compositeState, sqlExceptionNumber);
             shutdownSignal.Dispose();
-            Assert.Equal(4, activityProcessor.Invocations.Count); // OnStart/OnEnd/OnShutdown/Dispose called.
+            Assert.Equal(5, activityProcessor.Invocations.Count); // SetTracerProvider/OnStart/OnEnd/OnShutdown/Dispose called.
 
-            var activity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+            var activity = (Activity)activityProcessor.Invocations[2].Arguments[0];
 
             VerifyActivityData(commandType, commandText, captureText, isFailure, "127.0.0.1", activity, enableConnectionLevelAttributes);
         }
@@ -150,7 +150,7 @@ namespace OpenTelemetry.Instrumentation.SqlClient.Tests
 
             shutdownSignal.Dispose();
 
-            Assert.Equal(2, activityProcessor.Invocations.Count); // OnShutdown/Dispose called.
+            Assert.Equal(3, activityProcessor.Invocations.Count); // SetTracerProvider/OnShutdown/Dispose called.
         }
 
         [Fact]
@@ -169,7 +169,7 @@ namespace OpenTelemetry.Instrumentation.SqlClient.Tests
             fakeSqlEventSource.WriteEndExecuteEvent("arg1", "arg2", "arg3", "arg4");
             shutdownSignal.Dispose();
 
-            Assert.Equal(2, activityProcessor.Invocations.Count); // OnShutdown/Dispose called.
+            Assert.Equal(3, activityProcessor.Invocations.Count); // SetTracerProvider/OnShutdown/Dispose called.
         }
 
         private static void VerifyActivityData(