+++ src/OpenTelemetry/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## Unreleased
 
+* Removed `GetResource` and `SetResource` `Activity` extension methods. Added
+  `GetResource` extension method on `BaseProvider`
+  ([#1463](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1463))
+* Added `ParentProvider` property on `BaseProcessor` and `BaseExporter` classes.
+  ([#1463](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1463))
+* `Resource` is no longer added to observed `Activity` objects as a
+  `CustomProperty`.
+  ([#1463](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1463))
+* Remove RentrantExportProcessor as it is not required by spec.
+
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
+* TracerProviderBuilder API changes
+  Renamed AddInstrumentation to AddDiagnosticSourceInstrumentation
+  and made internal.
+  Added AddInstrumentation
+  ([#1454](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1454))
+
+* DiagnosticSource subscription helper classes (DiagnosticSourceSubscriber,
+  ListenerHandler,PropertyFetcher) are made internal.
+
 ## 0.7.0-beta.1
 
 Released 2020-Oct-16